requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
supabase>=2.0.0

# Development Dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
